requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==4.0.0
python-dotenv==1.0.0
boto3==1.28.44
requests==2.31.0
python-jose==3.3.0
google-auth-oauthlib>=0.4.6
google-auth>=2.22.0
google-auth-httplib2>=0.1.0
yt-dlp>=2023.11.16
moviepy==1.0.3
gunicorn==21.2.0
